Move `Copy` and `Paste` methods to `Items` namespace
