+++ bore.go
@@ -1,7 +1,6 @@
 package bore
 
 import (
-	"context"
 	"errors"
 	"os"
 	"strings"
@@ -9,36 +8,32 @@ import (
 
 	"github.com/uptrace/bun"
 	"go.trulyao.dev/bore/v2/database"
-	"go.trulyao.dev/bore/v2/database/models"
 	"go.trulyao.dev/bore/v2/database/repository"
 	"go.trulyao.dev/bore/v2/pkg/clipboard"
 	"go.trulyao.dev/bore/v2/pkg/events"
-	"go.trulyao.dev/bore/v2/pkg/events/aggregate"
-	"go.trulyao.dev/bore/v2/pkg/events/payload"
-	"go.trulyao.dev/bore/v2/pkg/lib"
-	"go.trulyao.dev/bore/v2/pkg/mimetype"
 )
 
-type (
-	Bore struct {
-		namespaceMutex sync.Mutex
+type Bore struct {
+	namespaceMutex sync.Mutex
 
-		// connection is the database connection used by this bore instance
-		db *bun.DB
+	// connection is the database connection used by this bore instance
+	db *bun.DB
 
-		// config holds the configuration for this bore instance
-		config *Config
+	// config holds the configuration for this bore instance
+	config *Config
 
-		// clipboard is the native clipboard interface for the current platform
-		clipboard clipboard.NativeClipboard
+	// clipboard is the native clipboard interface for the current platform
+	clipboard clipboard.NativeClipboard
 
-		// events is the event manager for this bore instance
-		manager *events.Manager
+	// events is the event manager for this bore instance
+	manager *events.Manager
 
-		// repository is the interface for accessing database operations
-		repository repository.Repository
-	}
-)
+	// repository is the interface for accessing database operations
+	repository repository.Repository
+
+	// MARK: Namespaces
+	items *itemsNamespace
+}
 
 var (
 	ErrInvalidArgs         = errors.New("invalid arguments provided to New function")
@@ -111,130 +106,16 @@ func (b *Bore) Config() (*Config, error) {
 	return b.config, nil
 }
 
-type CopyOptions struct {
-	Passthrough  bool   // Whether to also copy to the system clipboard if available.
-	CollectionID string // Optional collection ID to associate with the copied item.
-	Mimetype     mimetype.MimeType
-}
+// Items returns the items namespace for managing clipboard items.
+func (b *Bore) Items() *itemsNamespace {
+	b.namespaceMutex.Lock()
+	defer b.namespaceMutex.Unlock()
 
-// Copy copies the provided data to the Bore instance.
-func (b *Bore) Copy(ctx context.Context, data []byte, opts CopyOptions) error {
-	forwardToSystemClipboard := b.config.ClipboardPassthrough || opts.Passthrough
-	if b.clipboard.Available() && forwardToSystemClipboard {
-		if err := b.clipboard.Write(ctx, data); err != nil {
-			return err
-		}
+	if b.items == nil {
+		b.items = &itemsNamespace{b}
 	}
 
-	hash := lib.ComputeChecksum(data)
-	existingItem, err := b.repository.Items().FindByHash(ctx, hash, opts.CollectionID)
-	if err != nil {
-		return errors.New("failed to check for existing item: " + err.Error())
-	}
-
-	var e *events.Event
-	if existingItem != nil {
-		var existingAgg aggregate.Aggregate
-		existingAgg, err = aggregate.NewWithID(aggregate.AggregateTypeItem, existingItem.ID)
-		if err != nil {
-			return errors.New("failed to create aggregate for existing item: " + err.Error())
-		}
-
-		e, err = events.New(existingAgg, &payload.BumpItem{})
-	} else {
-		e, err = events.NewWithGeneratedID(
-			aggregate.AggregateTypeItem,
-			&payload.CreateItem{
-				Content:      data,
-				Mimetype:     opts.Mimetype,
-				CollectionID: opts.CollectionID,
-			},
-		)
-	}
-	if err != nil {
-		return errors.New("failed to create copy event: " + err.Error())
-	}
-
-	if _, _, err = b.manager.Apply(ctx, e, events.AppendOptions{ExpectedVersion: -1}); err != nil {
-		return errors.New("failed to apply copy event: " + err.Error())
-	}
-
-	return nil
-}
-
-type PasteOptions struct {
-	ItemID              string // Optional item identifier to filter pasted items.
-	CollectionID        string // Optional collection ID to filter pasted items.
-	FromSystemClipboard bool   // Whether to paste from the system clipboard if available.
-	DeleteAfterPaste    bool   // Whether to delete the pasted item after pasting.
-	SkipCollectionCheck bool   // Whether to skip checking if the collection exists.
-}
-
-type PasteResult struct {
-	Content []byte
-	Item    *models.Item
-}
-
-// Paste retrieves the last copied data from the Bore instance.
-func (b *Bore) Paste(ctx context.Context, options PasteOptions) (PasteResult, error) {
-	if b.clipboard.Available() && options.FromSystemClipboard {
-		rawContent, err := b.clipboard.Read(ctx)
-		if err != nil {
-			return PasteResult{}, errors.New("failed to read from system clipboard: " + err.Error())
-		}
-
-		return PasteResult{Content: rawContent, Item: nil}, nil
-	}
-
-	options.CollectionID = strings.TrimSpace(options.CollectionID)
-	if options.CollectionID != "" && !options.SkipCollectionCheck {
-		exists, err := b.repository.Collections().Exists(ctx, options.CollectionID)
-		if err != nil {
-			return PasteResult{}, errors.New("failed to check collection existence: " + err.Error())
-		} else if !exists {
-			return PasteResult{}, errors.New("requested collection does not exist")
-		}
-	}
-
-	var (
-		item *models.Item
-		err  error
-	)
-
-	identifier := strings.TrimSpace(options.ItemID)
-	if identifier == "" {
-		item, err = b.repository.Items().FindLatest(ctx, options.CollectionID)
-	} else {
-		item, err = b.repository.Items().FindById(ctx, identifier, options.CollectionID)
-	}
-
-	if err != nil {
-		return PasteResult{}, errors.New("failed to find latest item: " + err.Error())
-	}
-
-	if item == nil {
-		return PasteResult{}, nil
-	}
-
-	if options.DeleteAfterPaste {
-		agg, err := aggregate.NewWithID(aggregate.AggregateTypeItem, item.ID)
-		if err != nil {
-			return PasteResult{}, errors.New(
-				"failed to create aggregate for deletion: " + err.Error(),
-			)
-		}
-
-		e, err := events.New(agg, &payload.DeleteItem{})
-		if err != nil {
-			return PasteResult{}, errors.New("failed to create delete event: " + err.Error())
-		}
-
-		if _, _, err = b.manager.Apply(ctx, e); err != nil {
-			return PasteResult{}, errors.New("failed to apply delete event: " + err.Error())
-		}
-	}
-
-	return PasteResult{Content: item.Content, Item: item}, nil
+	return b.items
 }
 
 func (b *Bore) Close() error {
